Add chat emotion panel 10

+++ app/src/main/java/com/hwl/beta/db/dao/ChatGroupMessageDao.java
@@ -38,8 +38,8 @@ public class ChatGroupMessageDao extends AbstractDao<ChatGroupMessage, Long> {
         public final static Property OriginalUrl = new Property(11, String.class, "originalUrl", false, "ORIGINAL_URL");
         public final static Property ImageHeight = new Property(12, int.class, "imageHeight", false, "IMAGE_HEIGHT");
         public final static Property ImageWidth = new Property(13, int.class, "imageWidth", false, "IMAGE_WIDTH");
-        public final static Property Size = new Property(14, long.class, "size", false, "SIZE");
-        public final static Property PlayTime = new Property(15, long.class, "playTime", false, "PLAY_TIME");
+        public final static Property Size = new Property(14, int.class, "size", false, "SIZE");
+        public final static Property PlayTime = new Property(15, int.class, "playTime", false, "PLAY_TIME");
         public final static Property SendStatus = new Property(16, int.class, "sendStatus", false, "SEND_STATUS");
         public final static Property SendTime = new Property(17, java.util.Date.class, "sendTime", false, "SEND_TIME");
     }
@@ -239,8 +239,8 @@ public class ChatGroupMessageDao extends AbstractDao<ChatGroupMessage, Long> {
             cursor.isNull(offset + 11) ? null : cursor.getString(offset + 11), // originalUrl
             cursor.getInt(offset + 12), // imageHeight
             cursor.getInt(offset + 13), // imageWidth
-            cursor.getLong(offset + 14), // size
-            cursor.getLong(offset + 15), // playTime
+            cursor.getInt(offset + 14), // size
+            cursor.getInt(offset + 15), // playTime
             cursor.getInt(offset + 16), // sendStatus
             cursor.isNull(offset + 17) ? null : new java.util.Date(cursor.getLong(offset + 17)) // sendTime
         );
@@ -263,8 +263,8 @@ public class ChatGroupMessageDao extends AbstractDao<ChatGroupMessage, Long> {
         entity.setOriginalUrl(cursor.isNull(offset + 11) ? null : cursor.getString(offset + 11));
         entity.setImageHeight(cursor.getInt(offset + 12));
         entity.setImageWidth(cursor.getInt(offset + 13));
-        entity.setSize(cursor.getLong(offset + 14));
-        entity.setPlayTime(cursor.getLong(offset + 15));
+        entity.setSize(cursor.getInt(offset + 14));
+        entity.setPlayTime(cursor.getInt(offset + 15));
         entity.setSendStatus(cursor.getInt(offset + 16));
         entity.setSendTime(cursor.isNull(offset + 17) ? null : new java.util.Date(cursor.getLong(offset + 17)));
      }
